backend/requirements.txt
anthropic>=0.25.0
openai>=1.0.0
reportlab==4.0.4
chromadb==0.4.15
sentence-transformers==2.2.2
requests==2.31.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
pydantic-settings==2.0.3
sqlalchemy==2.0.20
asyncpg==0.28.0
aiofiles==23.2.1
fastapi==0.103.0
uvicorn==0.23.2
redis==5.0.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
aiosqlite==0.19.0
